requirements.txt
alembic==1.10.4
async-timeout==4.0.3
black==23.9.1
blinker==1.6.2
click==8.1.3
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
Faker==19.10.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.4
Flask-Pydantic==0.11.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.3.0
psycopg2==2.9.6
pydantic==1.10.7
pytest==7.4.2
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
six==1.16.0
SQLAlchemy==1.4.18
tomli==2.0.1
typing_extensions==4.5.0
uWSGI==2.0.22
Werkzeug==2.3.4
